# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Subnet Analytics Tournaments API",
    "description": "# Subnet Analytics Tournaments API\n\nPattern detection & feature generation evaluation platform for the Bittensor Subnet.\n\n## Tournament Scoring (Three Pillars)\n\nMiners are evaluated on **three pillars**:\n\n### 1. Features (10%)\n- Output schema validation for `features.parquet`\n- Generation performance relative to baseline\n\n### 2. Patterns (55%)\n- **Synthetic Recall (30%)**: % of ground_truth addresses detected\n- **Pattern Precision (25%)**: Anti-cheat via flow tracing validation\n\n### 3. Novelty (25%) + Performance (10%)\n- **Novelty Discovery (25%)**: New valid patterns beyond ground_truth\n- **Pattern Performance (10%)**: Detection speed relative to baseline\n\n## Flow Tracing Validation\n\nAll patterns (synthetic and novelty) are validated by verifying transaction flows exist in `transfers.parquet`. This prevents gaming through fake patterns.\n\n## API Versioning\n\nCurrent version: `v1`\n\nAll endpoints use `/api/v1/analytics/` prefix.\n\n## Authentication\n\nThis is a **read-only public API**. No authentication required.",
    "version": "1.0.0",
    "contact": {
      "name": "Subnet Team",
      "url": "https://github.com/subnet"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/api/v1/analytics/stats/miners/{hotkey}": {
      "get": {
        "tags": ["Stats"],
        "summary": "Get Miner History",
        "description": "Retrieve historical performance data for a specific miner across all tournaments.\n\n**Use Cases:**\n- Display miner profile page\n- Track miner improvement over time\n- Analyze performance trends",
        "operationId": "get_miner_history",
        "parameters": [
          {
            "name": "hotkey",
            "in": "path",
            "required": true,
            "description": "Miner's hotkey (SS58 address)",
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9]{48}$",
              "example": "5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty"
            }
          },
          {
            "$ref": "#/components/parameters/SkipParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Miner performance history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinerHistoryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Miner not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/stats/epochs/{epoch_number}": {
      "get": {
        "tags": ["Stats"],
        "summary": "Get Epoch Statistics",
        "description": "Retrieve statistics for a specific epoch's tournament.\n\n**Use Cases:**\n- Compare epoch performance\n- Analyze historical data\n- Generate epoch reports",
        "operationId": "get_epoch_stats",
        "parameters": [
          {
            "name": "epoch_number",
            "in": "path",
            "required": true,
            "description": "Epoch number",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "example": 42
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Epoch statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpochStatsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Epoch not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
...
      "SkipParam": {
        "name": "skip",
        "in": "query",
        "required": false,
        "description": "Number of records to skip for pagination",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of records to return",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["detail"],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "TournamentResultResponse": {
        "type": "object",
        "description": "Aggregated tournament result for a miner",
        "required": ["id", "tournament_id", "hotkey", "uid", "final_score", "rank"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tournament_id": {
            "type": "string",
            "format": "uuid"
          },
          "hotkey": {
            "type": "string"
          },
          "uid": {
            "type": "integer"
          },
          "output_schema_validity_rate": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true,
            "description": "Percentage of runs passing schema validation"
          },
          "pattern_existence_rate": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true,
            "description": "Percentage of runs finding patterns"
          },
          "feature_performance_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          },
          "synthetic_recall_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          },
          "pattern_precision_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          },
          "novelty_discovery_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          },
          "pattern_performance_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "nullable": true
          },
          "total_runs": {
            "type": "integer",
            "nullable": true
          },
          "total_patterns_reported": {
            "type": "integer",
            "nullable": true
          },
          "total_synthetic_found": {
            "type": "integer",
            "nullable": true
          },
          "total_novelty_valid": {
            "type": "integer",
            "nullable": true
          },
          "total_novelty_invalid": {
            "type": "integer",
            "nullable": true
          },
          "final_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "rank": {
            "type": "integer",
            "minimum": 1
          },
          "beat_baseline": {
            "type": "boolean",
            "description": "Whether this miner beat the baseline"
          },
          "is_winner": {
            "type": "boolean",
            "description": "Whether this is the tournament winner"
          },
          "calculated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MinerHistoryResponse": {
        "type": "object",
        "required": ["hotkey", "results"],
        "properties": {
          "hotkey": {
            "type": "string"
          },
          "total_tournaments": {
            "type": "integer"
          },
          "total_wins": {
            "type": "integer"
          },
          "average_rank": {
            "type": "number"
          },
          "average_score": {
            "type": "number"
          },
          "best_rank": {
            "type": "integer"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournamentResultResponse"
            }
          }
        }
      },
...
      "EpochStatsResponse": {
        "type": "object",
        "required": ["epoch_number"],
        "properties": {
          "epoch_number": {
            "type": "integer"
          },
          "tournament_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "total_participants": {
            "type": "integer"
          },
          "average_score": {
            "type": "number"
          },
          "top_score": {
            "type": "number"
          },
          "winner_hotkey": {
            "type": "string",
            "nullable": true
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      }
    }
  }
}